Convert upper-right refresh button to Share
