+++ src/components/ShareButton.tsx
@@ -1,0 +1,41 @@
+import { CustomIcon } from '@src/components/CustomIcon'
+import Tooltip from '@src/components/Tooltip'
+import usePlatform from '@src/hooks/usePlatform'
+import { hotkeyDisplay } from '@src/lib/hotkeyWrapper'
+import { commandBarActor } from '@src/machines/commandBarMachine'
+import { useHotkeys } from 'react-hotkeys-hook'
+
+const shareHotkey = 'mod+alt+s'
+const onShareClick = () =>
+  commandBarActor.send({
+    type: 'Find and select command',
+    data: { name: 'share-file-link', groupId: 'code' },
+  })
+
+/** Share Zoo link button shown in the upper-right of the modeling view */
+export const ShareButton = () => {
+  const platform = usePlatform()
+  useHotkeys(shareHotkey, onShareClick, {
+    scopes: ['modeling'],
+  })
+
+  return (
+    <button
+      type="button"
+      onClick={onShareClick}
+      className="flex gap-1 items-center py-0 pl-0.5 pr-1.5 m-0 bg-chalkboard-10/80 dark:bg-chalkboard-100/50 hover:bg-chalkboard-10 dark:hover:bg-chalkboard-100 border border-solid active:border-primary"
+    >
+      <CustomIcon name="link" className="w-5 h-5" />
+      <span className="flex-1">Share</span>
+      <Tooltip
+        position="bottom-right"
+        contentClassName="max-w-none flex items-center gap-4"
+      >
+        <span className="flex-1">Share part via Zoo link</span>
+        <kbd className="hotkey text-xs capitalize">
+          {hotkeyDisplay(shareHotkey, platform)}
+        </kbd>
+      </Tooltip>
+    </button>
+  )
+}
